+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [0.2.0] - 28/07/2025
 ### Added
 - new class Visualizer3D and parent class BaseVisualizer
-- new 'reduced' parameter to Visualizer classes and private _set_embeddings function to handle embeddings use
+- new 'reduced' parameter to Visualizer classes and 'auto' value for red_method in plot functions to automatically use cached reduced embeddings
+- private _set_embeddings function to handle embeddings use
 - 4 options of 3D plots analogous to 2D versions
 - 4 new aesthetic themes: light2, dark2, light3, dark3
 - optimized available pretrained json structure